fixed bug with forms that created groups and events even if the url was not inputed

+++ backend/db/migrations/20230320190022-create-event.js
@@ -43,7 +43,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       price: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.INTEGER
       },
       startDate: {
         allowNull: false,
